requirements.txt
django
djangorestframework
django-rest-swagger==2.1.2